main.py
import os
import asyncio
from dotenv import load_dotenv
from colorama import init, Fore

# Load environment variables
load_dotenv()
init(autoreset=True)

class NinymAssistant:
    def __init__(self):
        print(f"{Fore.CYAN}[SYSTEM] Initializing Ninym Assistant...")
        self.running = True
        
    async def start(self):
        print(f"{Fore.GREEN}[SYSTEM] Ninym is online! Waiting for input...")
        
        # Main Loop placeholder
        while self.running:
            try:
                # 1. Listen (STT)
                # 2. Think (LLM)
                # 3. Speak (TTS) & Act (Avatar)
                await asyncio.sleep(1) 
            except KeyboardInterrupt:
                self.running = False
                print(f"{Fore.RED}[SYSTEM] Shutting down...")

if __name__ == "__main__":
    bot = NinymAssistant()
    try:
        asyncio.run(bot.start())
    except KeyboardInterrupt:
        pass
